apps/backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

# Database
tortoise-orm==0.20.0
aiosqlite==0.17.0

# Data processing and ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.4.0

# Model explainability
shap>=0.45.0

# Visualization
plotly>=5.18.0
matplotlib>=3.9.0
seaborn>=0.13.0

# Validation and serialization
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Rate limiting
slowapi>=0.1.9

# File handling
aiofiles>=23.2.1
python-magic>=0.4.27

# Environment and configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
pytest-cov>=4.1.0

# Development tools
black>=23.11.0
isort>=5.12.0
ruff>=0.1.6

# Monitoring and logging
structlog>=23.2.0
